+++ shelter/contrib/config/__init__.py
@@ -3,3 +3,5 @@ Extended classes which encapsulates configuration.
 """
 
 from shelter.contrib.config.iniconfig import IniConfig
+
+__all__ = ['IniConfig']