+++ tests/settings5.py
@@ -6,17 +6,10 @@ MANAGEMENT_COMMANDS = (
 )
 
 INTERFACES = {
-    'http': {
+    'http_both_tcp_and_unix': {
         'LISTEN': ':4443',
-        'PROCESSES': 12,
-        'URLS': 'tests.urls1.urls_http',
-    },
-    'fastrpc': {
-        'LISTEN': '192.168.1.0:4445',
-    },
-    'unix': {
         'UNIX_SOCKET': '/tmp/tornado.socket',
         'PROCESSES': 12,
-        'URLS': 'tests.urls1.urls_unix',
+        'URLS': 'tests.urls1.urls_http',
     },
 }